+++ onmt/modules/Embeddings.py
@@ -190,3 +190,5 @@ class Embeddings(nn.Module):
         aeq(emb_size, self.embedding_size)
 
         return emb
+
+
