+++ test/Options.py
@@ -1,0 +1,63 @@
+class Opt:
+    def __init__(self):
+        self.data = '/Users/ugan/AnacondaProjects/OpenNMT-py/data/demo'
+        self.batch_size = 64
+        self.valid_batch_size = 32
+        self.batch_type = "sents"
+        self.gpuid = []
+        self.label_smoothing = 0.0
+        self.copy_attn = False
+        self.word_vec_size = -1
+        self.src_word_vec_size = 500
+        self.tgt_word_vec_size = 500
+        self.layers = -1
+        self.enc_layers = 2
+        self.dec_layers = 2
+        self.brnn = False
+        self.encoder_type = 'rnn'
+        self.seed = -1
+        self.exp_host = ''
+        self.tensorboard = False
+        self.report_every = 50
+        self.truncated_decoder = 0
+        self.max_generator_batches = 32
+        self.normalization = 'sents'
+        self.accum_count = 1
+        self.epochs = 13
+        self.start_epoch = 1
+        self.start_checkpoint_at = 0
+        self.save_model = 'demo-model'
+        self.train_from = ''
+        self.optim = 'sgd'
+        self.learning_rate = 1.0
+        self.max_grad_norm = 5
+        self.learning_rate_decay = 0.5
+        self.start_decay_at = 8
+        self.adam_beta1 = 0.9
+        self.adam_beta2 = 0.999
+        self.adagrad_accumulator_init = 0
+        self.decay_method = ''
+        self.warmup_steps = 4000
+        self.rnn_size = 500
+        self.rnn_type = 'LSTM'
+        self.model_type = 'text'
+        self.position_encoding = False
+        self.feat_merge = 'concat'
+        self.feat_vec_exponent = 0.7
+        self.feat_vec_size = -1
+        self.dropout = 0.3
+        self.bridge= False
+        self.share_embeddings = False
+        self.decoder_type = 'rnn'
+        self.input_feed = 1
+        self.global_attention = 'general'
+        self.coverage_attn = False
+        self.context_gate = None
+        self.reuse_copy_attn = False
+        self.share_decoder_embeddings = False
+        self.param_init = 0.1
+        self.pre_word_vecs_enc = None
+        self.fix_word_vecs_enc = False
+        self.pre_word_vecs_dec = None
+        self.fix_word_vecs_dec = False
+        self.param_init_glorot = False
